move utility bar to component
